# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.ResolveReportRequest:
    properties:
      action:
        description: resolve, dismiss
        type: string
      note:
        type: string
      reason:
        type: string
    required:
    - action
    type: object
...
  auth.ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        description: Details is present only for VALIDATION_ERROR responses.
      message:
        example: Human-readable error description.
        type: string
    type: object
...
  db.AuditLog:
    properties:
      action:
        type: string
      admin_id:
        type: string
      admin_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      new_status:
        type: string
      note:
        type: string
      old_status:
        type: string
      reason:
        type: string
      target_id:
        type: string
      target_type:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: dev@flickey.com
    name: Flickey Dev Team
  description: Property-rental platform backend. Provides OTP-based authentication,
    multi-step listing wizard, media upload workflow, amenity catalogue, and an admin
    moderation panel.
  license:
    name: MIT
  termsOfService: http://flickey.com/terms/
  title: Flickey Go API
  version: "1.0"
paths:
  /admin/audit:
    get:
      description: Returns recent administrative audit log entries.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.AuditLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - Admin
  /admin/listings:
    get:
      description: Returns a list of listings, optionally filtered by status.
      parameters:
      - description: Listing status filter (pending, active, rejected, suspended,
          archived)
        enum:
        - pending
        - active
        - rejected
        - suspended
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of listings with host details
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all listings
      tags:
      - Admin
  /admin/listings/{id}:
    get:
      description: Returns full listing details for moderation.
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: LISTING_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get listing details
      tags:
      - Admin
  /admin/listings/{id}/moderate:
    post:
      consumes:
      - application/json
      description: Approve, reject, or suspend a listing. Transitions state and records
        an audit log.
      parameters:
      - description: Listing UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Moderation action (approve/reject/request_changes/suspend) and
          reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.ModerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: VALIDATION_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Moderate a listing
      tags:
      - Admin
  /admin/overview:
    get:
      description: Returns counts of total users, listings, pending verifications,
        active listings, etc.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.OverviewMetrics'
        "401":
          description: MISSING_TOKEN / INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: INSUFFICIENT_ROLE
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get admin overview metrics
      tags:
      - Admin
  /admin/reports:
    get:
      description: Returns user/listing reports optionally filtered by status.
      parameters:
      - description: Report status filter (pending, resolved, dismissed)
        enum:
        - pending
        - resolved
        - dismissed
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Report'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List reports
      tags:
      - Admin
  /admin/reports/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Marks a moderation report as resolved or dismissed, recording note
        and reason.
      parameters:
      - description: Report UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Resolution action (resolve/dismiss) and details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.ResolveReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Report'
        "400":
          description: VALIDATION_ERROR / INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve or dismiss a report
      tags:
      - Admin
  /admin/users:
    get:
      description: Returns a list of users, optionally filtered by role and status.
      parameters:
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
  /admin/users/{id}:
    get:
      description: Returns a user's details and their listings.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details
      tags:
      - Admin
  /admin/users/{id}/enforce:
    post:
      consumes:
      - application/json
      description: Applies a warning, temporary block, or ban to a user. Records an
        audit log.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Enforcement action and reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.EnforceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: VALIDATION_ERROR / INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enforce user restrictions
      tags:
      - Admin
  /admin/users/{id}/role:
    post:
      consumes:
      - application/json
      description: Changes a user's role (guest, host, admin, moderator, support).
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Target user role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.RoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: VALIDATION_ERROR / INVALID_ID
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - Admin
  /admin/verification:
    get:
      description: Returns verification requests optionally filtered by status.
      parameters:
      - description: Verification status filter (pending, approved, rejected, changes_requested)
        enum:
        - pending
        - approved
        - rejected
        - changes_requested
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.VerificationRequest'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List identity verification requests
      tags:
      - Admin
...
  /listings:
    get:
      description: Returns all currently published listings. No authentication required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/listings.ListingPublicSchema'
            type: array
        "500":
          description: INTERNAL_ERROR
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Browse published listings
      tags:
      - Listings
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: **Bearer {token}**'
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: HttpOnly refresh token cookie set by /auth/verify-otp and rotated
      by /auth/refresh.
    in: cookie
    name: refresh_token
    type: apiKey
swagger: "2.0"